refactor: configuration in imports
